// app/src/generated/models/Ben_ltkcarddatasModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ben_ltkcarddatasstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ben_ltkcarddatasstatecode = keyof typeof Ben_ltkcarddatasstatecode;
export const Ben_ltkcarddatasstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ben_ltkcarddatasstatuscode = keyof typeof Ben_ltkcarddatasstatuscode;

export interface Ben_ltkcarddatasBase {
  ben_boardid?: string;
  ben_cardid: string;
  ben_cardtype?: string;
  "ben_Instance@odata.bind"?: string;
  ben_ltkcarddataid: string;
  ben_name?: string;
  ben_outputjson?: string;
  ben_tilesvg?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Ben_ltkcarddatasstatecode;
  statuscode?: Ben_ltkcarddatasstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ben_ltkcarddatas extends Ben_ltkcarddatasBase {
  ben_instancename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  ben_instance?: object;
  _ben_instance_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}